product attributes seeders

+++ database/seeds/ProductSeeder.php
@@ -7,6 +7,7 @@ class ProductSeeder extends Seeder
 {
     protected $tableName = 'products';
     protected $oldTableName = 'cyizj_jshopping_products';
+    protected $oldCategoriesTable = 'cyizj_jshopping_products_to_categories';
     /**
      * Run the database seeds.
      *
@@ -16,6 +17,9 @@ class ProductSeeder extends Seeder
     {
         if (!Schema::hasTable($this->oldTableName)) {
             throw new Exception("Please, copy old product table '{$this->oldTableName}'");
+        }
+        if (!Schema::hasTable($this->oldCategoriesTable)) {
+            throw new Exception("Please, copy old product table '{$this->oldCategoriesTable}'");
         }
 
         Schema::dropIfExists($this->tableName);
@@ -28,13 +32,30 @@ class ProductSeeder extends Seeder
             $table->renameColumn('product_id', 'id');
 
             $table->uuid('slug')->unique()->nullable();
+            $table->foreignId('category_id')->index();
+            $table->foreignId('season_id')->index();
+            $table->foreignId('color_id')->index();
+            $table->foreignId('brand_id')->index();
+
+
+
+            $table->foreignId('category_id')->index();
+
+
             $table->timestamps();
             $table->softDeletes();
         });
-
-
+        
+        // дата создания и обновления
         DB::update("UPDATE {$this->tableName} SET created_at = product_date_added, updated_at = date_modify
             WHERE product_date_added <> '0000-00-00 00:00:00' AND date_modify  <> '0000-00-00 00:00:00'");
+
+        // категория
+        DB::update("UPDATE products p, {$this->oldCategoriesTable} c SET p.category_id=c.category_id WHERE p.id=c.product_id");
+        // сезон
+        DB::update("UPDATE products SET season_id=extra_field_7");
+        // производитель
+        DB::update("UPDATE products SET brand_id=product_manufacturer_id");
 
         Schema::table($this->tableName, function (Blueprint $table) {
             $table->dropColumn('parent_id');
@@ -49,6 +70,7 @@ class ProductSeeder extends Seeder
             $table->dropColumn('currency_id');
             $table->dropColumn('product_template');
             $table->dropColumn('product_url');
+            $table->dropColumn('image');
 
             $table->dropColumn('name_en-GB');
             $table->dropColumn('alias_en-GB');
@@ -59,18 +81,27 @@ class ProductSeeder extends Seeder
             $table->dropColumn('meta_keyword_en-GB');
         });
 
+        Schema::table($this->tableName, function (Blueprint $table) {
+            $table->renameColumn('extra_field_1', 'color_txt');
+            $table->renameColumn('extra_field_2', 'fabric_top_txt');
+            $table->renameColumn('extra_field_8', 'fabric_inner_txt');
+            $table->renameColumn('extra_field_9', 'fabric_insole_txt');
+            $table->renameColumn('extra_field_10', 'fabric_outsole_txt');
+            $table->renameColumn('extra_field_11', 'heel_txt');
+        });
+
         $this->call(ProductAttributeSeeder::class);
 
-        /*Schema::table($this->tableName, function (Blueprint $table) {
-            $table->dropColumn('KIT');
-            $table->dropColumn('PRODUCT_SIZE');
-            $table->dropColumn('COLOR');
-            $table->dropColumn('COLLECTION');
-            $table->dropColumn('FABRIC');
-            $table->dropColumn('STYLE');
-            $table->dropColumn('HEIGHT');
-            $table->dropColumn('OUTLET_REF');
-            $table->dropColumn('SALE_REF');
-        });*/
+        Schema::table($this->tableName, function (Blueprint $table) {
+            $table->dropColumn('extra_field_3'); // Коллекция
+            $table->dropColumn('extra_field_7'); // Сезон
+            $table->dropColumn('extra_field_12'); // Размер аксессуара
+            $table->dropColumn('extra_field_13'); // Цвет фильтра
+            $table->dropColumn('extra_field_14'); // Материал фильтра
+            $table->dropColumn('extra_field_15'); // Теги
+            $table->dropColumn('extra_field_16'); // Акция
+            $table->dropColumn('extra_field_16'); // Поднять
+            $table->dropColumn('extra_field_16'); // Рейтинг
+        });
     }
 }
